Fix non-continuous link groups
The RPM payloadReader attempts to gracefully detect and handle hard
links - however, it was relying on the actual data appearing directly
after a series of hard links to that data. In practice, however, not all
RPMs follow this format, and the actual data may appear far after when
it is defined.

This patch fixes this behaviour by correctly tracking all inodes seen so
far, similarly to how the Extract() method already works today.

Co-Authored-By: Ilya Dmitrichenko <errordeveloper@gmail.com>
Signed-off-by: Justin Chadwell <me@jedevc.com>